+++ src/posts/posts.module.ts
@@ -3,6 +3,7 @@ import { CommonModule } from '@angular/common';
 import { PostCardComponent } from './post-card/post-card.component';
 import { PostThumbnailComponent } from './post-thumbnail/post-thumbnail.component';
 import { AddPostComponent } from './add-post/add-post.component';
+import { CommentsComponent } from './comments/comments.component';
 
 // Angular Material
 import { MatButtonModule } from '@angular/material/button';
@@ -11,14 +12,17 @@ import {MatSidenavModule} from '@angular/material/sidenav';
 import {MatIconModule} from '@angular/material/icon';
 import {MatMenuModule} from '@angular/material/menu';
 import { PostViewComponent } from './post-view/post-view.component';
-
+import {MatFormFieldModule} from '@angular/material/form-field';
+import {MatInputModule} from '@angular/material/input';
+import {MatExpansionModule} from '@angular/material/expansion';
 
 @NgModule({
   declarations: [
     PostCardComponent,
     PostThumbnailComponent,
     AddPostComponent,
-    PostViewComponent
+    PostViewComponent,
+    CommentsComponent
   ],
   imports: [
     CommonModule,
@@ -26,12 +30,16 @@ import { PostViewComponent } from './post-view/post-view.component';
     MatButtonModule,
     MatCardModule,
     MatSidenavModule,
-    MatMenuModule
+    MatMenuModule,
+    MatFormFieldModule,
+    MatInputModule,
+    MatExpansionModule
   ],
   exports: [
     AddPostComponent,
     PostCardComponent,
-    PostThumbnailComponent
+    PostThumbnailComponent,
+    CommentsComponent
   ]
 
 })